feat: add folder for MCC2022 Track1

+++ src/main/java/de/uulm/sp/fmc/as4moco/solver/SolverResponse.java
@@ -2,5 +2,5 @@ package de.uulm.sp.fmc.as4moco.solver;
 
 import java.util.Optional;
 
-public record SolverResponse(SolverInterface solver, SolverStatusEnum status, Optional<Long> solution) {
+public record SolverResponse(SolverInterface solver, SolverStatusEnum status, Optional<Double> solution) {
 }
